+++ README.md
@@ -1,1 +1,3 @@
-# csis3175-proj
+# Rooftown
+
+This is the source repository for CSIS3175 (W23) Group 11 Project
