+++ app/build.gradle
@@ -3,11 +3,11 @@ plugins {
 }
 
 android {
-    namespace 'csis3175.proj'
+    namespace 'csis3175.w23.g11.rooftown'
     compileSdk 32
 
     defaultConfig {
-        applicationId "csis3175.proj"
+        applicationId "csis3175.w23.g11.rooftown"
         minSdk 21
         targetSdk 32
         versionCode 1
