+++ app/src/androidTest/java/csis3175/w23/g11/rooftown/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package csis3175.proj;
+package csis3175.w23.g11.rooftown;
 
 import android.content.Context;
