+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Project"
+rootProject.name = "Rooftown"
 include ':app'
